+++ src/pages/api/v1/company/[companyId]/account/[accountId]/index.ts
@@ -3,10 +3,9 @@ import { IAccount } from '@/interfaces/accounting_account';
 import { IResponseData } from '@/interfaces/response_data';
 import { STATUS_MESSAGE } from '@/constants/status_code';
 import { formatApiResponse, isParamNumeric } from '@/lib/utils/common';
-import prisma from '@/client';
-import type { Account } from '@prisma/client';
 import { checkAdmin } from '@/lib/utils/auth_check';
-import { PUBLIC_COMPANY_ID } from '@/constants/company';
+import { formatAccount } from '@/lib/utils/formatter/account.formatter';
+import { findFirstAccountInPrisma } from '@/lib/utils/repo/account.repo';
 
 export function formatParams(
   companyId: unknown,
@@ -28,51 +27,6 @@ export function formatParams(
   };
 }
 
-export async function findFirstAccountInPrisma(accountId: number, companyId: number) {
-  let account: Account | null;
-  try {
-    account = await prisma.account.findFirst({
-      where: {
-        id: accountId,
-        OR: [
-          {
-            companyId,
-          },
-          {
-            companyId: PUBLIC_COMPANY_ID,
-          },
-        ]
-      },
-    });
-  } catch (error) {
-    // Info (20240516 - Murky) - Debugging error
-    // eslint-disable-next-line no-console
-    console.error(error);
-    throw new Error(STATUS_MESSAGE.DATABASE_READ_FAILED_ERROR);
-  }
-
-  if (!account) {
-    throw new Error(STATUS_MESSAGE.RESOURCE_NOT_FOUND);
-  }
-
-  return account;
-}
-
-export function formatAccounts(account: Account): IAccount {
-  return {
-    id: account.id,
-    companyId: account.companyId,
-    system: account.system,
-    type: account.type,
-    liquidity: account.liquidity,
-    debit: account.debit,
-    code: account.code,
-    name: account.name,
-    createdAt: account.createdAt,
-    updatedAt: account.updatedAt,
-  };
-}
-
 export async function handleGetRequest(
   req: NextApiRequest,
   res: NextApiResponse<IResponseData<IAccount>>
@@ -82,7 +36,7 @@ export async function handleGetRequest(
   const { accountId } = req.query;
   const { accountIdNumber, companyIdNumber } = formatParams(companyId, accountId);
   const accountFromDb = await findFirstAccountInPrisma(accountIdNumber, companyIdNumber);
-  const account = formatAccounts(accountFromDb);
+  const account = accountFromDb ? formatAccount(accountFromDb) : ({} as IAccount);
 
   const { httpCode, result } = formatApiResponse<IAccount>(STATUS_MESSAGE.SUCCESS, account);
   return {
