+++ src/pages/api/v1/company/[companyId]/account/index.test.ts
@@ -36,6 +36,7 @@ const mockAccountsInDB: Account[] = [
       name: 'Sun Bank',
       createdAt: 1000000000,
       updatedAt: 1000000000,
+      level: 1,
     }
   ];
   const mockAccounts: IAccount[] = [{
